requirements.txt
# LangChain Core
langchain>=0.2.0
langchain-groq>=0.1.2

# LLM APIs
groq>=0.15.0
openai>=1.30.1

# Voice Input (STT)
speechrecognition>=3.8.1
pyaudio>=0.2.14
pydub>=0.25.1

# Vision Input (image processing)
pillow>=11.0.0

# Text-to-Speech
gtts>=2.5.4
elevenlabs>=1.50.3

# Gradio UI
gradio>=4.26.0

# Helpers
python-dotenv>=1.0.1

# Audio playback (platform-dependent)
ffmpeg-python>=0.2.0
